+++ pkg/xstrings/funcs.go
@@ -1,0 +1,9 @@
+package xstrings
+
+func CopyMap(m map[string]string) map[string]string {
+	n := make(map[string]string, len(m))
+	for k, v := range m {
+		n[k] = v
+	}
+	return n
+}