+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add configuration linter under `lint` command.
 - Add logs when generating config via CLI.
+- Reconcile Config CRs.
 
 ### Fixed